feat: Refactor LottoStore to accept Money object in sell method
- Replace Long parameter in LottoStore.sell with Money to enforce type-safe monetary input.
- Remove validation logic from LottoStore, delegating responsibility entirely to Money.
- Delete obsolete test cases for invalid raw amounts in LottoStoreTest.
- Update all usages and tests to construct Money before calling sell.
